requirements.txt
APScheduler==3.6.3
database==0.6.0
dbutils==3.0.3
Flask==2.0.2
Flask_APScheduler==1.12.4
flask_cors==4.0.0
ipwhois==1.2.0
mock==5.1.0
paramiko==2.11.0
psutil==5.9.2
PyMySQL==1.0.2
PyYAML==6.0.1
redis==5.0.0
requests==2.21.0
setuptools==40.6.2
tornado==6.1
Werkzeug==2.0.2
